requirements.txt
streamlit
pandas
requests
streamlit_autorefresh
smartapi-python
pyotp
openai